requirements.txt
# Core AI/ML Libraries
tensorflow==2.14.0
torch==2.7.1
torchvision==0.22.1
transformers==4.53.2
numpy==1.26.4
scikit-learn==1.3.0

# Data Processing
pandas==2.3.1
opencv-python==4.7.0.72
Pillow==9.5.0

# Visualization
matplotlib==3.7.1
seaborn==0.13.2

# Web Framework
streamlit==1.25.0

# PDF Generation
fpdf==1.7.2

# Additional Dependencies (Auto-installed with above packages)
huggingface-hub==0.33.4
safetensors==0.5.3